[Chore] categoryList Obj, CategoryItem Data class 를 common 패키지로 이동
