+++ app/src/main/java/com/kust/kustaurant/presentation/ui/home/HomeFragment.kt
@@ -22,6 +22,7 @@ import com.kust.kustaurant.R
 import com.kust.kustaurant.data.getAccessToken
 import com.kust.kustaurant.data.model.HomeResponse
 import com.kust.kustaurant.databinding.FragmentHomeBinding
+import com.kust.kustaurant.presentation.common.CategoryData
 import com.kust.kustaurant.presentation.ui.detail.DetailActivity
 import com.kust.kustaurant.presentation.ui.search.SearchActivity
 import com.kust.kustaurant.presentation.ui.tier.TierFragment
@@ -143,26 +144,7 @@ class HomeFragment : Fragment(), CategoryAdapter.CategoryItemClickListener {
     }
 
     private fun setupCategoryRV(){
-        val categoryList = listOf(
-            CategoryItem(R.drawable.img_category_all, "전체"),
-            CategoryItem(R.drawable.img_category_korea, "한식"),
-            CategoryItem(R.drawable.img_category_japan, "일식"),
-            CategoryItem(R.drawable.img_category_china, "중식"),
-            CategoryItem(R.drawable.img_category_western, "양식"),
-            CategoryItem(R.drawable.img_category_asian, "아시안"),
-            CategoryItem(R.drawable.img_category_meat, "고기"),
-            CategoryItem(R.drawable.img_category_seafood, "해산물"),
-            CategoryItem(R.drawable.img_category_chicken, "치킨"),
-            CategoryItem(R.drawable.img_category_hamburger_pizza, "햄버거/피자"),
-            CategoryItem(R.drawable.img_category_tteokbokki, "분식"),
-            CategoryItem(R.drawable.img_category_beer, "술집"),
-            CategoryItem(R.drawable.img_category_cafe, "카페/디저트"),
-            CategoryItem(R.drawable.img_category_bakery, "베이커리"),
-            CategoryItem(R.drawable.img_category_salad, "샐러드"),
-            CategoryItem(R.drawable.img_category_benefit, "제휴업체"),
-        )
-
-        categoryAdapter = CategoryAdapter(categoryList, this)
+        categoryAdapter = CategoryAdapter(CategoryData.categoryList, this)
         binding.categoryRecyclerView.apply {
             val spanCount = 4
             val spacing = 10
